Fixes type error

+++ examples/stress.rs
@@ -29,7 +29,7 @@ where
     let end = time::Instant::now();
 
     let duration = end.duration_since(start);
-    let runtime_secs = duration.as_secs() + (duration.subsec_nanos() as f64 / 1_000_000_000.0);
+    let runtime_secs = duration.as_secs() + (duration.subsec_nanos() as f64 / 1_000_000_000.0) as u64;
     (res, runtime_secs)
 }
